+++ test/e2e/commands/tests/base-command-options.ts
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type SoloLogger} from '../../../../src/core/logging/solo-logger.js';
+import {type ClusterReferences} from '../../../../src/types/index.js';
+import {type NamespaceName} from '../../../../src/types/namespace/namespace-name.js';
+
+export interface BaseCommandOptions {
+  readonly testName: string;
+  readonly testLogger: SoloLogger;
+  readonly clusterReferences: ClusterReferences;
+  readonly clusterReferenceNameArray: string[];
+  readonly contexts: string[];
+  readonly deployment: string;
+  readonly namespace: NamespaceName;
+  readonly testCacheDirectory: string;
+  readonly enableLocalBuildPathTesting: boolean;
+  readonly localBuildReleaseTag: string;
+  readonly localBuildPath: string;
+  readonly createdAccountIds: string[];
+}
